+++ cmd/hcltm/dashboard.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"fmt"
+	"html/template"
 	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"sort"
 	"strings"
-	"text/template"
 	"time"
 
 	"github.com/xntrik/hcltm/pkg/spec"
@@ -34,6 +34,7 @@ type DashboardCommand struct {
 	flagDashboardTemplate   string
 	flagThreatmodelTemplate string
 	flagDashboardFilename   string
+	flagDashboardHtml       bool
 }
 
 func (c *DashboardCommand) Help() string {
@@ -60,6 +61,8 @@ Options:
 
  -dashboard-filename=<filename>
 
+ -dashboard-html
+
  -threatmodel-template=<file>
 
 `
@@ -75,6 +78,7 @@ func (c *DashboardCommand) Run(args []string) int {
 	flagSet.StringVar(&c.flagThreatmodelTemplate, "threatmodel-template", "", "Template file to override the default threatmodel.md file(s)")
 	flagSet.BoolVar(&c.flagOverwrite, "overwrite", false, "Overwrite existing files in the outdir. Defaults to false")
 	flagSet.BoolVar(&c.flagNoDfd, "nodfd", false, "Do not include generated DFD images. Defaults to false")
+	flagSet.BoolVar(&c.flagDashboardHtml, "dashboard-html", false, "Instead of writing .md files, write .html files instead")
 	flagSet.Parse(args)
 
 	if c.flagConfig != "" {
@@ -102,6 +106,11 @@ func (c *DashboardCommand) Run(args []string) int {
 		fmt.Println(c.Help())
 		return 1
 	} else {
+
+		outExt := "md"
+		if c.flagDashboardHtml {
+			outExt = "html"
+		}
 
 		// Parse the dashboard-index template first before creating folders
 
@@ -181,23 +190,23 @@ func (c *DashboardCommand) Run(args []string) int {
 		// We use outfiles to generate a list of output files to validate whether
 		// we're overwriting them or not.
 		outfiles := []string{
-			fmt.Sprintf("%s/%s.md", c.flagOutDir, c.flagDashboardFilename),
+			fmt.Sprintf("%s/%s.%s", c.flagOutDir, c.flagDashboardFilename, outExt),
 		}
 
 		// Find all the .hcl files we're going to parse
-		HCLFiles := findHclFiles(flagSet.Args())
+		AllFiles := findAllFiles(flagSet.Args())
 
 		// Parse all the identified .hcl files - just to determine output files
-		for _, file := range HCLFiles {
+		for _, file := range AllFiles {
 			tmParser := spec.NewThreatmodelParser(c.specCfg)
-			err := tmParser.ParseHCLFile(file, false)
+			err := tmParser.ParseFile(file, false)
 			if err != nil {
 				fmt.Printf("Error parsing %s: %s\n", file, err)
 				return 1
 			}
 
 			for _, tm := range tmParser.GetWrapped().Threatmodels {
-				outfile := outfilePath(c.flagOutDir, tm.Name, file, ".md")
+				outfile := outfilePath(c.flagOutDir, tm.Name, file, fmt.Sprintf(".%s", outExt))
 
 				outfiles = append(outfiles, outfile)
 
@@ -226,9 +235,9 @@ func (c *DashboardCommand) Run(args []string) int {
 
 		tmList := []tmListEntryType{}
 
-		for _, file := range HCLFiles {
+		for _, file := range AllFiles {
 			tmParser := spec.NewThreatmodelParser(c.specCfg)
-			err := tmParser.ParseHCLFile(file, false)
+			err := tmParser.ParseFile(file, false)
 			if err != nil {
 				fmt.Printf("Error parsing %s: %s\n", file, err)
 				return 1
@@ -262,7 +271,7 @@ func (c *DashboardCommand) Run(args []string) int {
 					return 1
 				}
 
-				outfile := outfilePath(c.flagOutDir, tm.Name, file, ".md")
+				outfile := outfilePath(c.flagOutDir, tm.Name, file, fmt.Sprintf(".%s", outExt))
 
 				f, err := os.Create(outfile)
 				if err != nil {
@@ -336,7 +345,7 @@ func (c *DashboardCommand) Run(args []string) int {
 
 		// Now we create the dashboard-index file
 
-		f, err := os.Create(c.flagOutDir + fmt.Sprintf("/%s.md", c.flagDashboardFilename))
+		f, err := os.Create(c.flagOutDir + fmt.Sprintf("/%s.%s", c.flagDashboardFilename, outExt))
 		if err != nil {
 			fmt.Printf("Error creating dashboard file: %s\n", err)
 			return 1
@@ -349,7 +358,7 @@ func (c *DashboardCommand) Run(args []string) int {
 			return 1
 		}
 
-		fmt.Printf("Successfully wrote to '%s/%s.md'\n", c.flagOutDir, c.flagDashboardFilename)
+		fmt.Printf("Successfully wrote to '%s/%s.%s'\n", c.flagOutDir, c.flagDashboardFilename, outExt)
 
 	}