+++ cmd/hcltm/dfd.go
@@ -89,12 +89,12 @@ func (c *DfdCommand) Run(args []string) int {
 		outfiles := []string{}
 
 		// Find all the .hcl files we're going to parse
-		HCLFiles := findHclFiles(flagSet.Args())
+		AllFiles := findAllFiles(flagSet.Args())
 
 		// Parse all the identified .hcl files - just to determine output files
-		for _, file := range HCLFiles {
+		for _, file := range AllFiles {
 			tmParser := spec.NewThreatmodelParser(c.specCfg)
-			err := tmParser.ParseHCLFile(file, false)
+			err := tmParser.ParseFile(file, false)
 			if err != nil {
 				fmt.Printf("Error parsing %s: %s\n", file, err)
 				return 1
@@ -139,9 +139,9 @@ func (c *DfdCommand) Run(args []string) int {
 			}
 		}
 
-		for _, file := range HCLFiles {
+		for _, file := range AllFiles {
 			tmParser := spec.NewThreatmodelParser(c.specCfg)
-			err := tmParser.ParseHCLFile(file, false)
+			err := tmParser.ParseFile(file, false)
 			if err != nil {
 				fmt.Printf("Error parsing %s: %s\n", file, err)
 				return 1